Moved operations to another file, Renamed and refactored code for Insert and Upsert
